Moved FeedStreamTo and ConsumeStreamFrom into deamon package and mde them private
